backend/ocr/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
google-cloud-vision>=3.7.0
pydantic>=2.6.0
